Replace / rename googleClientJobStore to googleJobStore
