+++ src/toil/jobStores/googleJobStore.py
@@ -21,10 +21,9 @@ import base64
 from contextlib import contextmanager
 import hashlib
 import uuid
-import boto
-from boto.exception import GSResponseError, GSDataError
 import logging
 import time
+from google.cloud import storage, exceptions
 
 try:
     import cPickle as pickle
@@ -40,7 +39,6 @@ from toil.jobStores.abstractJobStore import (AbstractJobStore, NoSuchJobExceptio
                                              ConcurrentFileModificationException)
 from toil.jobStores.utils import WritablePipe, ReadablePipe
 from toil.jobGraph import JobGraph
-from toil.jobStores.aws.utils import SDBHelper
 log = logging.getLogger(__name__)
 
 GOOGLE_STORAGE = 'gs'
@@ -58,25 +56,6 @@ GOOGLE_STORAGE = 'gs'
 
 class GoogleJobStore(AbstractJobStore):
 
-    # BOTO WILL UPDATE HEADERS WITHOUT COPYING THEM FIRST. To enforce immutability & prevent
-    # this, we use getters that return copies of our original dictionaries. reported:
-    # https://github.com/boto/boto/issues/3517
-    @property
-    def encryptedHeaders(self):
-        return self._encryptedHeaders.copy()
-
-    @encryptedHeaders.setter
-    def encryptedHeaders(self, value):
-        self._encryptedHeaders = value
-
-    @property
-    def headerValues(self):
-        return self._headerValues.copy()
-
-    @headerValues.setter
-    def headerValues(self, value):
-        self._headerValues = value
-
     def __init__(self, locator):
         super(GoogleJobStore, self).__init__()
 
@@ -87,78 +66,59 @@ class GoogleJobStore(AbstractJobStore):
             namePrefix = locator
             projectID = None
 
-        #  create 2 buckets
         self.locator = locator
         self.projectID = projectID
         self.bucketName = namePrefix+"--toil"
         log.debug("Instantiating google jobStore with name: %s", self.bucketName)
-        self.gsBucketURL = bytes("gs://"+self.bucketName)
-        self._headerValues = {"x-goog-project-id": bytes(projectID)} if projectID else {}
-        self._encryptedHeaders = self.headerValues
 
-        import gcs_oauth2_boto_plugin # needed to import authentication handler
-        self.uri = boto.storage_uri(self.gsBucketURL, GOOGLE_STORAGE)
-        self.files = None
+        import gcs_oauth2_boto_plugin  # needed to import authentication handler
+        # this is a :class:`~google.cloud.storage.bucket.Bucket`
+        self.bucket = None
 
         self.statsBaseID = 'f16eef0c-b597-4b8b-9b0c-4d605b4f506c'
         self.statsReadPrefix = '_'
         self.readStatsBaseID = self.statsReadPrefix+self.statsBaseID
 
+        self.sseKey = None
+
     def initialize(self, config=None):
-        exists = True
+        storageClient = storage.Client()
         try:
-            self.uri.get_bucket(headers=self.headerValues, validate=True)
-        except GSResponseError:
-            exists = False
-
-        if exists:
+            self.bucket = storageClient.create_bucket(self.bucketName)
+        except exceptions.Conflict:
             raise JobStoreExistsException(self.locator)
-
-        self.files = self._retryCreateBucket(self.uri, self.headerValues)
-
-        # functionally equivalent to dictionary1.update(dictionary2) but works with our immutable dicts
-        self.encryptedHeaders = dict(self.encryptedHeaders, **self._resolveEncryptionHeaders(config))
-
         super(GoogleJobStore, self).initialize(config)
 
+        # set up sever side encryption after we set up config in super
+        if self.config.sseKey is not None:
+            with open(self.config.sseKey) as f:
+                self.sseKey = bytes(f.read())
+                assert len(self.sseKey) == 32
+
     def resume(self):
+        storage_client = storage.Client()
         try:
-            self.files = self.uri.get_bucket(headers=self.headerValues, validate=True)
-        except GSResponseError:
+            self.bucket = storage_client.get_bucket(self.bucketName)
+        except exceptions.NotFound:
             raise NoSuchJobStoreException(self.locator)
         super(GoogleJobStore, self).resume()
 
-    @property
-    def sseKeyPath(self):
-        return self.config.sseKey
-
     def destroy(self):
-        # no upper time limit on this call keep trying delete calls until we succeed - we can
-        # fail because of eventual consistency in 2 ways: 1) skipping unlisted objects in bucket
-        # that are meant to be deleted 2) listing of ghost objects when trying to delete bucket
-
-        # just return if not connect to physical storage
-        if self.files is None:
+        # just return if not connect to physical storage. Needed for idempotency
+        if self.bucket is None:
             return
-        while True:
-            try:
-                self.uri.delete_bucket()
-            except GSResponseError as e:
-                if e.status == 404:
-                    return  # the bucket doesn't exist so we are done
-                else:
-                    # bucket could still have objects, or contain ghost objects
-                    time.sleep(0.5)
-            else:
-                # we have succesfully deleted bucket
-                return
 
-            # object could have been deleted already
-            for obj in self.files.list():
-                try:
-                    obj.delete()
-                except GSResponseError:
-                    pass
+        try:
+            self.bucket.delete(force=True)
+            # throws ValueError if bucket has more than 256 objects. Then we must delete manually
+        except ValueError:
+            self.bucket.delete_blobs(self.bucket.list_blobs)
+            self.bucket.delete()
+            # if ^ throws a google.cloud.exceptions.Conflict, then we should have a deletion retry mechanism.
+
+        # google freaks out if we call delete multiple times on the bucket obj, so after success
+        # just set to None.
+        self.bucket = None
 
     def create(self, jobNode):
         jobStoreID = self._newJobID()
@@ -168,163 +128,153 @@ class GoogleJobStore(AbstractJobStore):
         if hasattr(self, "_batchedJobGraphs") and self._batchedJobGraphs is not None:
             self._batchedJobGraphs.append(job)
         else:
-            self._writeString(jobStoreID, pickle.dumps(job, protocol=pickle.HIGHEST_PROTOCOL)) #UPDATE: bz2.compress(
+            self._writeString(jobStoreID, pickle.dumps(job, protocol=pickle.HIGHEST_PROTOCOL))  # UPDATE: bz2.compress(
         return job
 
+    # TODO: abstract?
     def _newJobID(self):
         return "job"+str(uuid.uuid4())
 
     def exists(self, jobStoreID):
-        # used on job files, which will be encrypted if avaliable
-        headers = self.encryptedHeaders
-        try:
-            self._getKey(jobStoreID, headers)
-        except NoSuchFileException:
-            return False
-        return True
+        return self.bucket.blob(bytes(jobStoreID), encryption_key=self.sseKey).exists()
 
     def getPublicUrl(self, fileName):
-        try:
-            key = self._getKey(fileName)
-        except:
+        blob = self.bucket.get_blob(bytes(fileName), encryption_key=self.sseKey)
+        if blob is None:
             raise NoSuchFileException(fileName)
-
-        #https://github.com/GoogleCloudPlatform/gcs-oauth2-boto-plugin/issues/15
-        # The google connection is missing sign_string().
-        # It is only signing 'GET', though, so skipping the encoding might be fine.
-        def sign_string(self, string_to_sign):
-            #import base64
-            #from oauth2client.crypt import Signer
-            #signer = Signer.from_string(self.oauth2_client._private_key)
-            #return base64.b64encode(signer.sign(string_to_sign))
-            return string_to_sign
-        import types
-        key.bucket.connection._auth_handler.sign_string = types.MethodType(sign_string, key.bucket.connection._auth_handler)
-        key.set_canned_acl('public-read')
-        return key.generate_url(query_auth=False,
-                               expires_in=self.publicUrlExpiration.total_seconds())
+        return blob.generate_signed_url(self.publicUrlExpiration)
 
     def getSharedPublicUrl(self, sharedFileName):
         return self.getPublicUrl(sharedFileName)
 
+    # TODO: abstract?
     def load(self, jobStoreID):
         try:
             jobString = self._readContents(jobStoreID)
         except NoSuchFileException:
             raise NoSuchJobException(jobStoreID)
-        return pickle.loads(jobString) #UPDATE bz2.decompress(
+        return pickle.loads(jobString)  # UPDATE bz2.decompress(
 
+    # TODO: abstract?
     def update(self, job):
         self._writeString(job.jobStoreID, pickle.dumps(job, protocol=pickle.HIGHEST_PROTOCOL), update=True)
 
+    # TODO: abstract? probably cannot, unless abstract listing files
     def delete(self, jobStoreID):
-        # jobs will always be encrypted when avaliable
-        self._delete(jobStoreID, encrypt=True)
+        self._delete(jobStoreID)
+
+        # best effort delete associated files
+        for blob in self.bucket.list_blobs(prefix=bytes(jobStoreID)):
+            self._delete(blob.name)
 
     def jobs(self):
-        for key in self.files.list():
-            jobStoreID = key.name
-            if len(jobStoreID) == 39: # 'job' + uuid length
+        for blob in self.bucket.list_blobs(prefix=b'job'):
+            jobStoreID = blob.name
+            if len(jobStoreID) == 39:  # 'job' + uuid length
                 yield self.load(jobStoreID)
 
+    # TODO: abstract?
     def writeFile(self, localFilePath, jobStoreID=None):
         fileID = self._newID(isFile=True, jobStoreID=jobStoreID)
         with open(localFilePath) as f:
             self._writeFile(fileID, f)
         return fileID
 
+    # TODO: abstract?
     @contextmanager
     def writeFileStream(self, jobStoreID=None):
         fileID = self._newID(isFile=True, jobStoreID=jobStoreID)
-        key = self._newKey(fileID)
-        with self._uploadStream(key, update=False) as writable:
-            yield writable, key.name
+        with self._uploadStream(fileID, update=True) as writable:
+            yield writable, fileID
 
+    # TODO: abstract?
     def getEmptyFileStoreID(self, jobStoreID=None):
         fileID = self._newID(isFile=True, jobStoreID=jobStoreID)
         self._writeFile(fileID, StringIO(""))
         return fileID
 
     def readFile(self, jobStoreFileID, localFilePath):
-        # used on non-shared files which will be encrypted if avaliable
-        headers = self.encryptedHeaders
-        # checking for JobStoreID existance
-        if not self.exists(jobStoreFileID):
+        # used on non-shared files which will be encrypted if available
+        # TODO deal with encryption stuff
+        # checking for JobStoreID existence
+        if not self.fileExists(jobStoreFileID):
             raise NoSuchFileException(jobStoreFileID)
         with open(localFilePath, 'w') as writeable:
-            self._getKey(jobStoreFileID, headers).get_contents_to_file(writeable, headers=headers)
+            blob = self.bucket.get_blob(bytes(jobStoreFileID), encryption_key=self.sseKey)
+            blob.download_to_file(writeable)
 
+    # TODO abstract?
     @contextmanager
     def readFileStream(self, jobStoreFileID):
         with self.readSharedFileStream(jobStoreFileID, isProtected=True) as readable:
             yield readable
 
+    # TODO abstract?
     def deleteFile(self, jobStoreFileID):
-        headers = self.encryptedHeaders
-        try:
-            self._getKey(jobStoreFileID, headers).delete(headers)
-        except GSDataError as e:
-            # we tried to delete unencrypted file with encryption headers. unfortunately,
-            # we can't determine whether the file passed in is encrypted or not beforehand.
-            if e.status == 400:
-                headers = self.headerValues
-                self._getKey(jobStoreFileID, headers).delete(headers)
-            else:
-                raise e
+        self._delete(jobStoreFileID)
 
     def fileExists(self, jobStoreFileID):
-        try:
-            self._getKey(jobStoreFileID)
-            return True
-        except (NoSuchFileException, GSResponseError) as e:
-            if isinstance(e, NoSuchFileException):
-                return False
-            elif e.status == 400:
-                # will happen w/ self.fileExists(encryptedFile). If file didn't exist code == 404
-                return True
-            else:
-                return False
+        return self.bucket.blob(bytes(jobStoreFileID), encryption_key=self.sseKey).exists()
 
     def updateFile(self, jobStoreFileID, localFilePath):
         with open(localFilePath) as f:
             self._writeFile(jobStoreFileID, f, update=True)
 
+    # TODO abstract?
     @contextmanager
     def updateFileStream(self, jobStoreFileID):
-        headers = self.encryptedHeaders
-        key = self._getKey(jobStoreFileID, headers)
-        with self._uploadStream(key, update=True) as readable:
-            yield readable
+        with self._uploadStream(jobStoreFileID, update=True) as writable:
+            yield writable
 
+    # TODO abstract?
     @contextmanager
     def writeSharedFileStream(self, sharedFileName, isProtected=True):
-        key = self._newKey(sharedFileName)
-        with self._uploadStream(key, encrypt=isProtected, update=True) as readable:
-            yield readable
+        with self._uploadStream(sharedFileName, encrypt=isProtected, update=True) as writable:
+            yield writable
 
+    # TODO abstract?
     @contextmanager
     def readSharedFileStream(self, sharedFileName, isProtected=True):
-        headers = self.encryptedHeaders if isProtected else self.headerValues
-        key = self._getKey(sharedFileName, headers=headers)
-        with self._downloadStream(key, encrypt=isProtected) as readable:
+        with self._downloadStream(sharedFileName, encrypt=isProtected) as readable:
             yield readable
 
-    @staticmethod
-    def _getResources(url):
-        return 'gs://' + url.netloc + url.path
+    @classmethod
+    def _getBlobFromURL(cls, url, exists=False):
+        """
+        Gets the blob specified by the url.
+
+        caution: makes no api request. blob may not ACTUALLY exist
+
+        :param urlparse.ParseResult url: the URL
+
+        :param bool exists: if True, then syncs local blob object with cloud
+        and raises exceptions if it doesn't exist remotely
+
+        :return: the blob requested
+        :rtype: :class:`~google.cloud.storage.blob.Blob`
+        """
+        bucketName = url.netloc
+        fileName = url.path
+
+        storageClient = storage.Client()
+        bucket = storageClient.get_bucket(bucketName)
+        blob = bucket.blob(bytes(fileName))
+
+        if exists:
+            if not blob.exists():
+                raise NoSuchFileException
+            # sync with cloud so info like size is available
+            blob.reload()
+        return blob
 
     @classmethod
     def getSize(cls, url):
-        uri = GoogleJobStore._getResources(url)
-        uri = boto.storage_uri(uri, GOOGLE_STORAGE)
-        return uri.get_key().size
+        return cls._getBlobFromURL(url, exists=True).size
 
     @classmethod
     def _readFromUrl(cls, url, writable):
-        # gs://projectid/bucket/key
-        uri = GoogleJobStore._getResources(url)
-        uri = boto.storage_uri(uri, GOOGLE_STORAGE)
-        uri.get_contents_to_file(writable)
+        blob = cls._getBlobFromURL(url, exists=True)
+        blob.download_to_file(writable)
 
     @classmethod
     def _supportsUrl(cls, url, export=False):
@@ -332,17 +282,16 @@ class GoogleJobStore(AbstractJobStore):
 
     @classmethod
     def _writeToUrl(cls, readable, url):
-        uri = GoogleJobStore._getResources(url)
-        uri = boto.storage_uri(uri, GOOGLE_STORAGE)
-        uri.set_contents_from_stream(readable)
+        blob = cls._getBlobFromURL(url)
+        blob.upload_from_file(readable)
 
     def writeStatsAndLogging(self, statsAndLoggingString):
         statsID = self.statsBaseID + str(uuid.uuid4())
-        key = self._newKey(statsID)
-        log.debug("Writing stats file: %s", key.name)
-        with self._uploadStream(key, encrypt=False, update=False) as f:
+        log.debug("Writing stats file: %s", statsID)
+        with self._uploadStream(statsID, encrypt=False, update=False) as f:
             f.write(statsAndLoggingString)
 
+    # TODO: abstract (only possible if list and blob.name are abstracted)
     def readStatsAndLogging(self, callback, readAll=False):
         prefix = self.readStatsBaseID if readAll else self.statsBaseID
         filesRead = 0
@@ -351,22 +300,21 @@ class GoogleJobStore(AbstractJobStore):
         while True:
             filesReadThisLoop = 0
             # prefix seems broken
-            for key in list(self.files.list()): #prefix=prefix)):
-                if not key.name.startswith(prefix):
-                    continue
+            for blob in self.bucket.list_blobs(prefix=bytes(prefix)):
                 try:
-                    with self.readSharedFileStream(key.name) as readable:
-                        log.debug("Reading stats file: %s", key.name)
+                    with self.readSharedFileStream(blob.name) as readable:
+                        log.debug("Reading stats file: %s", blob.name)
                         callback(readable)
                         filesReadThisLoop += 1
                     if not readAll:
                         # rename this file by copying it and deleting the old version to avoid
                         # rereading it
-                        newID = self.readStatsBaseID + key.name[len(self.statsBaseID):]
-                        self.files.copy_key(newID, self.files.name, key.name)
-                        key.delete()
+                        newID = self.readStatsBaseID + blob.name[len(self.statsBaseID):]
+                        # NOTE: just copies then deletes old.
+                        # TODO: abstract renaming for ultimate efficiency
+                        self.bucket.rename_blob(blob, bytes(newID))
                 except NoSuchFileException:
-                    log.debug("Stats file not found: %s", key.name)
+                    log.debug("Stats file not found: %s", blob.name)
             if readAll:
                 # The readAll parameter is only by the toil stats util after the completion of the
                 # pipeline. Assume that this means the bucket is in a consistent state when readAll
@@ -390,19 +338,6 @@ class GoogleJobStore(AbstractJobStore):
         return filesRead
 
     @staticmethod
-    def _retryCreateBucket(uri, headers):
-        # FIMXE: This should use retry from utils
-        while True:
-            try:
-                # FIMXE: this leaks a connection on exceptions
-                return uri.create_bucket(headers=headers)
-            except GSResponseError as e:
-                if e.status == 429:
-                    time.sleep(10)
-                else:
-                    raise
-
-    @staticmethod
     def _newID(isFile=False, jobStoreID=None):
         if isFile and jobStoreID:  # file associated with job
             return jobStoreID+str(uuid.uuid4())
@@ -411,144 +346,92 @@ class GoogleJobStore(AbstractJobStore):
         else:  # job id
             return "job"+str(uuid.uuid4())
 
-    def _resolveEncryptionHeaders(self, config):
-        sseKeyPath = config.sseKey
-        if sseKeyPath is None:
-            return {}
-        else:
-            with open(sseKeyPath) as f:
-                sseKey = f.read()
-            assert len(sseKey) == 32
-            encodedSseKey = base64.b64encode(sseKey)
-            encodedSseKeyMd5 = base64.b64encode(hashlib.sha256(sseKey).digest())
-            return {'x-goog-encryption-algorithm': 'AES256',
-                    'x-goog-encryption-key': encodedSseKey,
-                    'x-goog-encryption-key-sha256': encodedSseKeyMd5,
-                    "Cache-Control": "no-store"}
+    # TODO: abstract and require implementation?
+    def _delete(self, jobStoreFileID):
+        if self.fileExists(jobStoreFileID):
+            self.bucket.get_blob(bytes(jobStoreFileID)).delete()
+        # remember, this is supposed to be idempotent, so we don't do anything
+        # if the file doesn't exist
 
-    def _delete(self, jobStoreID, encrypt=True):
-        headers = self.encryptedHeaders if encrypt else self.headerValues
-        try:
-            key = self._getKey(jobStoreID, headers)
-        except NoSuchFileException:
-            pass
-        else:
-            try:
-                key.delete()
-            except GSResponseError as e:
-                if e.status == 404:
-                    pass
-        # best effort delete associated files
-        for fileID in self.files.list():
-            if fileID.name.startswith(jobStoreID):
-                try:
-                    self.deleteFile(fileID)
-                except NoSuchFileException:
-                    pass
-
-    def _getKey(self, jobStoreID=None, headers=None):
-
-        # UPDATE _getKey() has problems - listing buckets works though
-        for obj in self.uri.get_bucket():
-            #print '%s://%s/%s' % (self.uri.scheme, self.uri.bucket_name, obj.name)
-            if jobStoreID == obj.name:
-                return obj
-
-        # gets remote key, in contrast to self._newKey
-        key = None
-        try:
-            key = self.files.get_key(jobStoreID, headers=headers)
-        except GSDataError:
-            if headers == self.encryptedHeaders:
-                # https://github.com/boto/boto/issues/3518
-                # see self._writeFile for more
-                pass
-            else:
-                raise
-        if key is None:
-            raise NoSuchFileException(jobStoreID)
-        else:
-            return key
-
-    def _newKey(self, jobStoreID):
-        # Does not create a key remotely. Careful -- does not ensure key name is unique
-        return self.files.new_key(jobStoreID)
-
+    # TODO: abstract and require implementation?
     def _readContents(self, jobStoreID):
-        # used on job files only, which will be encrypted if avaliable
-        headers = self.encryptedHeaders
-        return self._getKey(jobStoreID, headers).get_contents_as_string(headers=headers)
+        """
+        To be used on files representing jobs only. Which will be encrypted if possible.
+        :param jobStoreID: the ID of the job
+        :type jobStoreID: str
+        :return: contents of the job file
+        :rtype: string
+        """
+        job = self.bucket.get_blob(bytes(jobStoreID), encryption_key=self.sseKey)
+        if job is None:
+            raise NoSuchJobException(jobStoreID)
+        return job.download_as_string()
 
+    # TODO: abstract and require implementation?
     def _writeFile(self, jobStoreID, fileObj, update=False, encrypt=True):
-        headers = self.encryptedHeaders if encrypt else self.headerValues
-        if update:
-            key = self._getKey(jobStoreID=jobStoreID, headers=headers)
+        blob = self.bucket.blob(bytes(jobStoreID), encryption_key=self.sseKey if encrypt else None)
+        if not update:
+            # TODO: should probably raise a special exception and be added to all jobStores
+            assert not blob.exists()
         else:
-            key = self._newKey(jobStoreID=jobStoreID)
-        headers = self.encryptedHeaders if encrypt else self.headerValues # UPDATE: why is this duplicated?
-        try:
-             key.set_contents_from_file(fileObj, headers=headers)
-        except GSDataError:
-            if encrypt:
-                # Per https://cloud.google.com/storage/docs/encryption#customer-supplied_encryption_keys
-                # the etag and md5 will not match with customer supplied
-                # keys. However boto didn't get the memo apparently, and will raise this error if
-                # they dont match. Reported: https://github.com/boto/boto/issues/3518
-                pass
-            else:
-                raise
+            if not blob.exists():
+                raise NoSuchFileException(jobStoreID)
+        blob.upload_from_file(fileObj)
 
+    # TODO: abstract?
     def _writeString(self, jobStoreID, stringToUpload, **kwarg):
         self._writeFile(jobStoreID, StringIO(stringToUpload), **kwarg)
 
     @contextmanager
-    def _uploadStream(self, key, update=False, encrypt=True):
-        store = self
+    def _uploadStream(self, fileName, update=False, encrypt=True):
+        """
+        Yields a context manager that can be used to write to the bucket
+        with a stream. See :class:`~toil.jobStores.utils.WritablePipe` for an example.
+
+        Will throw assertion error if the file shouldn't be updated
+        and yet exists.
+
+        :param fileName: name of file to be inserted into bucket
+        :type fileName: str
+        :param update: whether or not the file is to be updated
+        :type update: bool
+        :param encrypt: whether or not the file is encrypted
+        :type encrypt: bool
+        :return: an instance of WritablePipe.
+        :rtype: :class:`~toil.jobStores.utils.writablePipe`
+        """
+        blob = self.bucket.blob(bytes(fileName), encryption_key=self.sseKey if encrypt else None)
 
         class UploadPipe(WritablePipe):
             def readFrom(self, readable):
-                headers = store.encryptedHeaders if encrypt else store.headerValues
-                if update:
-                    try:
-                        key.set_contents_from_stream(readable, headers=headers)
-                    except GSDataError:
-                        if encrypt:
-                            # https://github.com/boto/boto/issues/3518
-                            # see self._writeFile for more
-                            pass
-                        else:
-                            raise
-                else:
-                    try:
-                        # The if_generation argument insures that the existing key matches the
-                        # given generation, i.e. version, before modifying anything. Passing a
-                        # generation of 0 insures that the key does not exist remotely.
-                        key.set_contents_from_stream(readable, headers=headers, if_generation=0)
-                    except (GSResponseError, GSDataError) as e:
-                        if isinstance(e, GSResponseError):
-                            if e.status == 412:
-                                raise ConcurrentFileModificationException(key.name)
-                            else:
-                                raise e
-                        elif encrypt:
-                            # https://github.com/boto/boto/issues/3518
-                            # see self._writeFile for more
-                            pass
-                        else:
-                            raise
+                if not update:
+                    assert not blob.exists()
+                blob.upload_from_file(readable)
 
         with UploadPipe() as writable:
             yield writable
 
     @contextmanager
-    def _downloadStream(self, key, encrypt=True):
-        store = self
+    def _downloadStream(self, fileName, encrypt=True):
+        """
+        Yields a context manager that can be used to read from the bucket
+        with a stream. See :class:`~toil.jobStores.utils.WritablePipe` for an example.
+
+        :param fileName: name of file in bucket to be read
+        :type fileName: str
+        :param encrypt: whether or not the file is encrypted
+        :type encrypt: bool
+        :return: an instance of ReadablePipe.
+        :rtype: :class:`~toil.jobStores.utils.ReadablePipe`
+        """
+        blob = self.bucket.get_blob(bytes(fileName), encryption_key=self.sseKey if encrypt else None)
+        if blob is None:
+            raise NoSuchFileException(fileName)
 
         class DownloadPipe(ReadablePipe):
             def writeTo(self, writable):
-                headers = store.encryptedHeaders if encrypt else store.headerValues
                 try:
-                    key.get_file(writable, headers=headers)
+                    blob.download_to_file(writable)
                 finally:
                     writable.close()
